+++ src/Contracts/ReverseGeocodingServiceContract.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dominasys\FilamentLocation\Contracts;
+
+use Dominasys\FilamentLocation\Data\GooglePlaceData;
+
+interface ReverseGeocodingServiceContract
+{
+    public function reverse(float $latitude, float $longitude): GooglePlaceData;
+}